Extract terminal settings section
